+++ src/main/java/coffeandcommit/crema/domain/review/dto/response/ReviewResponseDTO.java
@@ -6,6 +6,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -18,7 +19,7 @@ public class ReviewResponseDTO {
 
     private Long reviewId;         // 리뷰 ID
     private Long reservationId;    // 예약 ID
-    private float starReview;        // 별점
+    private BigDecimal starReview;        // 별점
     private String comment;        // 후기 내용
     private LocalDateTime createdAt; // 작성 시간
 
@@ -28,7 +29,7 @@ public class ReviewResponseDTO {
         return ReviewResponseDTO.builder()
                 .reviewId(review.getId())
                 .reservationId(review.getReservation().getId())
-                .starReview(review.getStarReview().floatValue())
+                .starReview(review.getStarReview())
                 .comment(review.getComment())
                 .createdAt(review.getCreatedAt())
                 .experienceEvaluations(
